refactor(discord-api): modularize interactions route handlers
Split the Discord interactions API route into focused modules for signature verification, command handling, component/modal handling, and shared interaction types while preserving endpoint behavior.
